Implement `Sweep` for `Edge`
