+++ crates/fj-kernel/src/algorithms/sweep.rs
@@ -1,5 +1,6 @@
 //! Sweeping objects along a path to create new objects
 
+mod edge;
 mod face;
 mod sketch;
